pyproject.toml
[tool.poetry]
name = "aeris"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Jay"]

[tool.poetry.scripts]
lint = 'black . && isort .'

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.2.1"
pytz = "^2023.3"
types-pytz = "^2023.3.0.1"
colorama = "^0.4.6"
icalendar = "^5.0.7"
types-colorama = "^0.4.15.12"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.5.1"
isort = "^5.12.0"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
